backend/pyproject.toml
[project]
name = "lms"
version = "0.1.0"
description = "LMS - Learning Managment System"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "alembic==1.18.4",
  "fastapi[standard]==0.135.1",
  "platformdirs==4.9.4",
  "psycopg[binary]==3.3.3",
  "pwdlib[argon2]==0.3.0",
  "pydantic-settings>=2.13.1",
  "pyjwt==2.11.0",
  "sqlalchemy[asyncio]==2.0.48",
  "structlog==25.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.3.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
